tests: Fix memory leak in Free callback
The DestroyInstanceAllocationCallbacksCompatibility test leaking
memory when run under windows is intentional due to the nature of
memory allocation in windows. Because the intent of this change is to
enable leak sanitizer runs for linux, the windows test leaking is not
an issue. While it is not desireable, there is no simple solution for
it because the test is 'doing bad behavior' in order to make sure
VVL catches the bad behavior.