ci: Enable Address Sanitizer
This commit enabled Address Sanitizer in github actions CI runs. This
requires enabling ASAN in the loader with the option to disable
unloading of dynamic libraries to enable useful stack traces.

Because leaks are unavoidable in certain circumstances, a mechanism
has been setup to allow tests to disable leak detection.
The tests which currently need this are:
* VkLayerTest.LeakASwapchain - intentionally leaks memory to trigger
    the test case
* VkSyncValTest.SyncQSPresentAcquire - known leak with fix in process

To disable leak checking in a test, add the following to the start
of a test.

```
#if defined(VVL_ENABLE_ASAN
    auto leak_sanitizer_disabler = __lsan::ScopedDisabler();
#endif
```

__lsan::ScopedDisabler() is a RAII type that comes from the
<sanitizer/lsan_interface.h> header, which is automatically
included if VVL_ENABLE_ASAN is defined.