tests: Fix external memory tests for swiftshader
With current spec (1.3.242) the validation of VU 00656 can be done only
in vkBindMemory. With invalid input (e.g. handle type is not supported
for any resource) vkAllocateMemory can legally fail. This does not give
a chance for validation to run. This fix detects vkAllocateMemory
failure and skips the test since validation is not possible.