layer: Alternative layer settings
Provide alternative validation layer settings from "enables" and
"disables" flags. The new settings can be parsed and interpreted
directly by vkconfig.

The new settings are converted into the current data structure that
remains unchanged with this PR but will be updated in a future PR.

The following steps are:
- Using the new LayerUtils library to load settings
- Refactoring the layer settings code to convert the flags into
the new settings.