layers: Improve an error message for vkMapMemory
When the client attempts to map a memory object without HOST_VISIBLE_BIT
we now print the memory type index and the property flags.  We also drop
the manual printing of the memory object handle since that should happen
automatically.  For the multi-instance check, we drop most of the
message because the layers will already print out the VUID text
automatically so it's redundant.