layers: Fix UB for VK_IMAGE_TILING_DRM_FORMAT_MODIFIER_EXT
Several codepaths in the layers called into
vkGetPhysicalDeviceImageFormatProperties2KHR without including a valid
VkPhysicalDeviceImageDrmFormatModifierInfoEXT, despite this being a
requirement for images with DRM format modifier tiling.

For the vkBindImageMemory check, there is a hypothetical possibility of
the driver returning VK_ERROR_OUT_OF_HOST_MEMORY when querying the
image's DRM format. In this case, for lack of anything better to do, we
preserve the current behavior.

It might make more sense to query the DRM format modifier immediately
after image creation and cache it inside the image state, but I didn't
want to make that (more invasive) change myself.

As an aside, we also correctly specify the specified `sharingMode` when
querying for DRM modifier support, which the current code failed to do
properly.

Co-authored-by: Simon Ser <contact@emersion.fr>

closes #5649
closes #5687