layers: Report api names for both race participants
Currently, when the race happens, VVL reports api name from the
current thread but does not say which call was on the other side
of the race.

Call stack might be misleading: another thread is not blocked and
can progress a bit further during short period of time that current
thread takes to reach the breakpoint (can be reproduced).