layers: Ignore pNext structs in GPLs
VkPipelineRenderingCreateInfo is part of fragment output state, and thus
should be ignored when a pipeline is being created without fragment
output state _except_ for the viewMask field.

This change adds a "copy state" object that is passed down the pNext
chain to potentially modify any pNext "safe copy". This could be used to
remove more "manual" code that is currently defined inside of scripts,
but for now this only affects VkPipelineRenderingCreateInfo.