layers: Remove global stage ordering functions
These functions can produce unexpected result. The problem is
that GetLogicallyEarLiest(Latest)GraphicsPipelineStage tries to create
global ordering of all pipeline stages (not only graphics as the name
suggests). So, if the mask specifies fragment+raytracing stages then
the earliest stage will always be evaluated to a fragment stage,
independenty if the application logic is concerned with the fragment or
with the raytracing stage.