tests: Improve testing via vvl_utils
Move the majority of validation layer code into the vvl_utils
static library. This allows the tests to link against vvl_utils
and test as much code as possible.

The actual vvl target should be as minimal as possible.

The only thing our vvl target requires is exporting specific
functions for the vulkan loader.