layers: Apply latest sync XML
The commit that implemented new syncval codegen used the first version
of the sync XML data. That data was not complete but it matched the
original syncval codegen input, so it helped to minimize diff.

This change re-generates syncval code using the final sync XML data
(from 1.3.253) which includes additional changes/fixes to fully reflect
the spec.