layers: Update logic to handle syncpipeline without TOP/BOTTOM
TOP/BOTTOM stages were removed from sync XML but syncval generator still
needs to generate the same output which puts TOP/BOTTOM in some places.

The plan is to update generator, so it does not use TOP and BOTTOM
stages in the output, but it's for a separate task since additional code
changes/testing are needed.

Keep this as a separate commit, so it will be easier later to re-work
and undo this functionality.