layers: Do not check self-dependency in ValidateDependencyInfo
ValidateRenderPassPipelineBarriers is used to validate subpass
self-dependencies. It can be called through these code paths:
1. By ValidateCmdPipelineBarrier2
2. By ValidateCmdSetEvent2/ValidateCmdWaitEvents2

In the first case we have two duplicate calls to
ValidateRenderPassPipelineBarriers. It's called directly by
ValidateCmdPipelineBarrier2 but also as part of ValidateDependencyInfo.

In the second case a VkEvent is not allowed to form a self-dependency
inside a subpass (according to spec). Because the code did that anyway
it resulted in false-positive error messages.

The solution is to remove ValidateRenderPassPipelineBarriers call from
ValidateDependencyInfo.

With this fix the self-dependency validation is done by
PreCallValidateCmdPipelineBarrier and ValidateCmdPipelineBarrier2
that directly invoke ValidateRenderPassPipelineBarriers.