cmake: Fix MSVC warning 4459
The main issue is the globals instance_extensions and
device_extensions defined in chassis.cpp

This name is already used by ValidationObject which is what is
causing the warning.

This commit also removes the usage of ARRAY_SIZE in favor of
std::array and removes extern "C" usage that didn't need to exist.