scripts: Restore the -o option on generate_source
The old script had a way to specify a custom output directory
for generated files. The recent refactoring removed this option.

This option is important if your CI runs on a read-only version
of the repository, and generated files/intermediates are written to
a different folder before being processed (Bazel-like build).

Signed-off-by: Nathan Gauër <brioche@google.com>