layers: Check out of bounds access in syncval
This can happen when core checks are disabled or when corresponding
checks not implemented.

There are core validation checks for set-not-bound error for
vkCmdBindDescriptorSets API. Corresponding functionality is not
implemented yet for vkCmdBindDescriptorBuffers +
vkCmdSetDescriptorBufferOffsets API. That is why, even with core
checks enabled, we still need to do preventing checks in SyncVal.