layers: Check for unhandled vuids
It's similar to other GetVUID helpers to handle situation when VUID is
not defined.

Can happen when VU is defined only for Sync2 but we still do the checks
for Sync1 (e.g. VUID-VkMemoryBarrier2-srcAccessMask-06256). This
assumes that rules for compatible subset of Sync2 enums make sense for
Sync1 too.