gpu: Move GPU-AV helper functions to VVL
This change moves much of the GPU-AV instrumentation logic from
SPIRV-Tools to this code base. It also allows that logic to be
written in GLSL instead of generated SPIRV. This is intended
to make it much simpler to improve GPU-AV and to reduce the
frequency of breaking changes where VVL and SPIRV-Tools need
to be updated in lock step.

Note: this change does require specific versions of glslang and
SPRIV-Tools