layers: Use vvl::Func in SemOp and reorder members
Reduces size of semaphore state structures on 64 bit platforms.
The numbers is for the Release build, Debug build can add additional
debug state.

SemOp    : 40  -> 32 bytes
TimePoint: 104 -> 96 bytes

That's preparation for the next commit which will increase size of
TimePoint in order to use build-in storage of small_vector to avoid
allocation for common case of single semaphore wait. The goal is to
reduce size enough in order not to go over 2 cache lines (128 bytes)
in the next commit.

The size of SemOp won't be changed by the follow up commit, which is
nice - in the case of multiple SemOps we can put 2 ops per cache line.