layers: Use small_vector instead of std::set for SemOp waits
This avoids allocation for a common case of a semaphore with a single
wait (always for a binary semaphore and in practice often for timeline
semaphores - apps don't always go wild with queues).

The previous commit allowed to reduce sizes of SemOp/Timeline
structures and here we borrow that size back. TimePoint structure
in Release build on x64 is 128 bytes (perfectly fits 2 cache lines).
The size of SemOp remains 32 bytes.

The original code did not use properties of set container, only
iteration over all elements, so replacing with array is straightfoward.