docs: Correct SPIR-V compilation instructions
The documentation for debug printf includes a couple
example shaders and a short example command for how to
compile them. These commands include the -x flag which
exports the code in text-based hexadecimal format,
rather than the binary SPIR-V representation, which
might be confusing for someone looking to learn from
the document.