layers: Check that timepoint's wait ops use the same payload
It should be true in order to safely use small_vector
instead of std::set in TimePoint. Code inspection
suggets that it's true. TimePoint, that corresponds to
specific payload, collects wait ops with the same payload.
Added asserts to additionally validate this assumption.
Potentially we can simplify further if that's true.