layers: Introduce barrier structures
New types are similar to sync2 types with a benefit of hierarchical
data organization. For example, BufferBarrier can also be used by the
apis that accept QueueFamilyBarrier or MemoryBarrier types.

New barrier types can be initialized with both sync2 and sync1 types.
For sync1 types, stage masks should be additionally provided, because
they are not part of the sync1 barrier structures.

sType/pNext are not included. Currently it's not needed for validation
and if there is a need the idea to pass that information using higher
level concepts or separately.