layers: Checks that counters do not go negative
Negative value is a bug. In addition to indication it's a mismatch
between Start/Finish it's also leads to broken subsequent validation:
the next "Start" operation will not produce positive value and
that read/write won't be taken into account against other parallel
reads/writes.

The plan is to use unsigned values for counters after testing this
code for a bit.