layers: Remove ValidateDeviceObject
This function is no-op after chassis is introduced.

The chassis retrieves ValidationObject using a key derived from the
dispatchable handle. This guarantees that there can't be a mismatch
between ValidationObject and the first dispatchable parameter.

If the dispatchable handle is invalid, this will be detected
early by the chassis code (most likely a crash during dereferencing).

`ValidateDeviceObject` was added in the period when each validation
object was a separate layer. For configurations where ObjectLifetimes
was the first layer in the chain, `ValidateDeviceObject` had early
opportunity to scan through dispatchable object map before
`get_dispatch_key` is called. With the current design it's the chassis
that calls `get_dispatch_key`. If original 2017 behavior is desired,
then the validity and reporting of the first dispatchable handle
should be done at the chassis level.