build: Use vk_struct_helper.hpp instead
Replaces usage of vk_typemap_helper.h with vk_struct_helper.hpp from
the Vulkan-Utility-Libraries. It has the same functionality but since
this helper library lives in multiple places, it was moved to
Vulkan-Utility-Libraries to reduce code duplication.

Many places are able to use the vku::InitStructHelper class to deduce
the type to be initialized. This helper class does not work with auto
nor with a variadic argument list.