tests: Process single element in descriptor write helpers
The semantics of these functions was a bit confusing. In Vulkan, array
count means that multiple resources are consumed, and in our helpers,
array count means that a single resource will be duplicated this number
of times.

Updated helpers work only with a single element, which is a majority
of cases.

If in the future, we will need to specify multiple elements, then new
overloads can be added, with semantics more similar to Vulkan - element
count and pointer to array of resources.