layers: Autogen enabled feature detection
This change adds autogeneration for the DeviceFeatures struct as well as
a helper to populate it.

The DeviceFeatures struct is now a flat list of features; i.e. members
of VkPhysicalDevice*Features* classes.  This facilitates looking up
features everywhere else as it is irrelevant through which struct the
feature was enabled.  It will also reduce churn the next time a struct
such as VkPhysicalDeviceVulkan13Features is introduced.

Paired with this struct is a helper that populates the DeviceFeatures
struct with the contents of VkPhysicalDeviceFeatures as well as whatever
feature struct is found in the pNext chain of the device create info.