tests: Cleanup tests and help Thread Sanitizer
The (not)hidden agenda behind this change is to simplify life
for the Thread Sanitizer (TS). It can't recognize that queue mutex is
private to each QUEUE_STATE object and produces false-positive.

If I'm wrong in any of the following considerations lets assume it's
just a test cleanup (command pools owns command bufferes and
FreeCommandBuffers is optional).

TS reports possible deadlock due to inverted pair of locks:
(Q-mutex -> CmdBuf-mutex) and (CmdBuf->mutex -> Q-mutex)
but in our situation we can only have:
(Q1-mutex -> CmdBuf-mutex) and (CmdBuf->mutex -> Q2-mutex)
which can't lead to deadlock.

Also TS made the above analysis based only a single queue thread
(the tests of interest run only single queue thread), which could not
lead to a deadlock by definition.

The presense of vk::FreeCommandBuffers at the end of the test in rare
cases (not sure why on CI it became more often) leads to a situation
that command buffer is destroyed by the "Queue Thread" instead of
by the "Shutdown sequence of VkRenderFramework". During "Queue Thread"
destruction we have QueueLock->CmdBufLock that confuses TS. When we
move CB destruction to "Shutdown sequence of VkRenderFramework" then it
guarantees that QueueLock is released and only CmdBufLock is active
during command buffer destruction. This should make TS happy.