layers: Mark color attachment used when external resolve present
VK_ANDROID_external_format allows for the only color attachment
to be VK_ATTACHMENT_UNUSED. This lead to the blend state not
being copied for the safe_structs used by VVL. If run with
GPUAV or shaderprint, the struct sent to the driver will not
contain a blend state when the one from the application does.
This change fixes that issue.