gpu: Refacto validation and err msg logic
Adding validation for a command that needs to track the lifetime of its
specific validation resources is a bit painful, as you need to update
"Uber" functions, and classes, mainly
AllocateValidationResources, GenerateValidationMessage,
and the gpuav::CommandBuffer class.

The new CommandResources class aims at improving upon that.
It stores validation resources common to all recorded commands,
and if some command also needs specific resources, it can derive
from it. Objects of those new classes are stored as CommandResources
objects, there custom Destroy and error logging functions will be
automatically called.

Further work is needed, to improve the AnalyzeAndGenerateMessages
and GenerateValidationMessage functions for instance