gpu: Don't wait for idle after every submit
Add virtual methods to the CommandBuffer and Queue classes that
get run at PreRecord, PostRecord and Retire time. This makes it
possible to remove the PostRecord wait for idle from GPU-AV
submissions. Instead, at timeline semaphore is set up for host
waits in the Retire() virtual method so that GPU-AV post processing
for each submission is appropriately synchronized.