gpu: Workaround for multiple variables using a single binding
For bindless, it is common for multiple variables with different
types (such as 2d and 3d images) to be attached to a single
binding.  Avoid false positives in this case by looking for a
compatible variable definition to validate against rather than
just using the first one found.

A better fix would be to have the spirv instrumentation record
the variable id that produced each access and then use that
variable definiton for validation. However that will need to be
done later.