sync: Fix WaitForFences leak
There are syncval endpoints that compute state in the Validate phase
and use thread local storage to propagate it to the Record stage,
which does the final update.

This change fixes one side effect of this approach in the multithreaded
environment.

In following scenario two threads work with the same queue.
The queue access is externally synchronized.

1. Thread A: Submits some memory accesses. Signals a fence.
2. Thread B: Submits empty command buffer. Finishes only Validate.
   This imports acccesses from step 1 and the batch is stored in the
   thread local storage to be consumed later by the Record.
3. Thread A: Calls WaitForFences with a fence from step 1.
   This processes accesses from step 1, but can not see the copy
   of those accesses (in TLS) produced by step 2.
4. Thread B. Finishes Record. This restores accesses from step 2, that
   otherwise should have been processed by the WaitForFences.
5. Thread A. Submits more accesses. They hazard with accesses that
   were restored in step 4.