sync: Fix loop and stop using uint8_t as size type
The typo in the loop was not discovered because after running for 256
iterations the loop terminated due to uint8_t wrap around. Using
uint32_t makes similar issues more noticeable.

We always had this typo. Not sure which scenarios this could affect,
but as the first step this should be fixed.

Not the first issue with uint8_t we encountered, so start using
uint32_t as default size type for small_vector. This does not affect
the final size of the container due to alignment gaps. For specific
case of small_vector<ResourceFirstAccess, 3> it's 96 bytes both for
uint8_t and uint32_t size type parameterization.