layers: Add debug label tracking VU 01912
The validation of labels in the primary command buffer happens on the
queue level because debug regions can cross command buffer boundaries.
Validation should be done at submit time.

This is more complicated comparing to a similar VU for a secondary
command buffer where debug regions are not allowed to cross cross
command buffer boundaries.