tests: Add async hazard tests
The positive test is a regression test from:
https://github.com/KhronosGroup/Vulkan-ValidationLayers/issues/7386.

The QSWriteRacingWrite/QSWriteRacingWrite2 negative tests were added to
increase WRITE-RACING-WRITE coverage, but they do not reproduce the
issue that was fixed.

The QSWriteRacingRead test is a regression test for the initial
incorrect attempt to fix the issue (using scalar async tag tracking
instead of per-queue), which nevertheless did not cause any failures.