sync: Handle command without debug label and simplify code
If there is no debug label for a specific position in the command
stream, we use Uint32Max as a label command index. We need to check
this case and skip debug region reporting. It was done in one place
but was missing in another place.

Adding necessary check allowed to simplify implementation. Two
implementations of GetDebugRegionName became almost identical
(in CommandBuffer and QueueBatch context). The main logic was moved
into a shared helper function.