layers: Rework DescriptorSet invalidation
In StateObject::NotifyInvalidate(), the unlink parameter was only
ever set to true when a descriptor in a DescriptorSet became invalid.
But for bindless, no invalidation of state happened at all. This was
because invalidating an unused descriptor is not an error in bindless.

However, a CommandBuffer might be tracking layout state for an Image
descriptor inside its image_layout_map. When an Image is destroyed,
this state needs to be cleaned up. Otherwise its possible to end
up with destroyed memory in the map.

Instead add CommandBuffer::NotifyUpdateDescriptor() which can more
flexibly be used to tell the command buffer what to do.

Fixes #7435