layers: Add a unique object id to StateObject
Drivers which reuses handles can confuse some parts of validation,
such as per-CB image layout state tracking. Add an id that is unique
even when handles or state object pointers might not work. This is
somewhat redundant with Handle Wrapping but is available even when
that feature is turned off.