layers: Store Image Ids in CB ImageSubresourceLayoutMaps
In bindless applications it is possible for an image to be destroyed
while a command buffer has cached layout state for it in this map.
If the driver then reuses the image handle value for a new image,
layout validation can be confused or crash. Use the object id
to prevent this.