layers: Do not store acquire command in timeline SemOps
For acquire operation we don't need queue/seq. It can be represented
in a more compact form separately.

In some places this makes logic more specific. No need to check
acquire if we need to signal something on the queue (TimePoint::Notify).
But in general it's one more check in addition to signal. This is not
necessarily bad - this makes acquire paths more explicit and we can
think more about them.

That's one more step towards storing in timeline only references to
submissions.