layers: Update seq default value in SubmissionReference
The initial idea was to use kU64Max to refer to the entire queue's
workload (similar to how we do this in the Notify API). But it was
never used this way in the PresentSync path, from what I can observe.

Since SubmissionReference is used now in other places, it makes sense
to use zero as default seq value. This will match the previous way to
enqueue operations that do not happen on the queue (null queue, zero
seq).

Also renamed GetLastBinarySignalSubmission ->
GetPendingBinarySignalSubmission. "Last" can include operations that
already retired (as in LastOp).