gpu: Update to address validation
Bug fix: early exit in gpuav::Validator::UpdateBuffer would skip
table cleanup
Improvements:
- Remove unnecessary "copy to vector" step when getting device
addresses, instead directly copy to GPU memory
- Redid GPU buffer device address table management:
It is now a 1 to 1 mapping with what is held on the CPU,
which IMO simplify things
- Rework validation shader, it was hard to understand

Tests:
- Basic test is not working, there is at least an issue in how
buffer address accesses length
are computed. Access size seems to always be 8,
or the size of a pointer, and not the size of the type pointed to.