sync: Add SignaledSemaphoresUpdate
This splits SignaledSemaphores into two parts.
The part that is repsonsible for collecting updates is put into
a separate type - SignaledSemaphoresUpdate. The pars that is left
that's the new SignaledSemaphores and is just a map.

This simplifies design and restricts possible operations that
can be done with each type since they are used in different
contexts and have different interfaces.