sync: Refactor batch tag range initialization
After this change QueueBatchContext::tag_range_ has a single and
independent from the context semantics - global range of tags.

Previously, depending on the context, the tag_range was either a local
range (just a tag count, begin field was always zero) or a global
range (after all initialization was done). This was confusing at times
because you need to check the context of the code.

As a bonus reduced LOC and few functions removed from the interface.