sync: Rename indexing operator to GetAccessRecord
The query is a two-level search. Indexing operator does not provide a
hint what's going on and can mislead that it is a vector/map access.

Also this operation is used only in one place, so not a scenario of a
common operation that can benefit from a simpler syntax.