gpu: Fix a race condition in submission processing
Move most submission post processing into the PostCallRecord hooks.
This avoids adding submission state in the PreCall hook and then looking
for it in the PostCall hook. It also avoids adding bad state in case
the submit call fails.

The remaining work in the PreCall hook is setting up resources used by
the instrumented shaders we're about to execute and that must happen
before the submit is dispatched.