layers: Track present fences
It's the second attempt after:
https://github.com/KhronosGroup/Vulkan-ValidationLayers/pull/8400

Original present fence synchronization was implemented here:
https://github.com/KhronosGroup/Vulkan-ValidationLayers/pull/7694.
It was based on PresentSync, although PresentSync was not designed for
this (only for acquire-based synchronization).

The angle issue happened because we accidentally ended up with two
solutions handling the present fence: QueueSubmission and
PresentSync. This solution removes code path that uses PresentSync
and uses QueueSubmission fence instead.