layers: Break dependency of in-use tracking on GPU state
The idea is that in-use tracking is a not a GPU side system. Evaluate
resource in-use status based on CPU side API usage pattern. Make
GPU/driver errors a separate axis of error handling.

The problem with mixing these two types of errors handling is that it
leads to confusion in practice. For example, if application is correctly
structured to handle resource lifetimes or image layouts transition,
but then due to poor handling of VkResult the error message reports
that resource is still in use (but we just called vkDeviceWaitIdle!) or
layout is incorrect when all transitions are properly specified.
This misleading messages can focus attention on the wrong area.

If neccesary we might provide an option to detect non successful
VkResult checks (have this already?) for critical places like
submit/wait calls.

For now this change is experimental to observe how it works in practice.