layers: Fix and cleanup ownership transfer code
This removes ValidatorState ownership transfer helper.
It was mostly additional abstraction layer that copied local variables
and the result was stored as another local variable.

Also replaced VUID-vkQueueSubmit-pSubmits-04626 with unassigned VUID
when resource is EXCLUSIVE. 04626 is only about checking CONCURRENT
resource properties. It also checks resource itself but not a barrier
properties which is needed in unassigned VUID.