layers: Change ownership of AcquireFenceSync object
AcquireNextImage does not necessarily use the same fence with a
specific swapchain image. The previous implementation made this
(incorrect) assumption, and this allowed to use Fence object as the
owner of AcquireFenceSync.

A better solution is that Present stores AcquireFenceSync in a
swapchain image. When swapchain image is re-acquired, it assigns
AcquireFenceSync to a Fence object, if any.

This also removes the need to handle a special case that swapchain can
be re-created, and in that case we had to invalidate AcquireFenceSync
stored inside Fence. Now Fence object is automatically associated
with correct swapchain.