sync: Remove access entries of destroyed resources
Syncval cleanups old access entries during host side synchronization
operations. For example, when frame synchronization uses vkWaitForFences
to wait for the QueueSubmit's fence, this guarantees that device memory
accesses are completed and it is safe to remove corresponding tracking
state.

There are synchronization mechanisms that are not associated directly
with a queue (acquire fence sync) and do not guarantee completion
of memory accesses (provide only execution dependency). In this case
we need to cleanup access entries when resource is destroyed, because
there is no guarantee they will be removed later.