sync: Introduce key-value properties section
This section is optional and is disabled by default. If enabled it is
appended to the main error messages.

The purpose of the key-values section is to allow filtering of the
error messages based on selected properties. This section can also
include debugging information.

This aims to solve a problem when the applications extract parts of the
main error message for filtering purposes (false positives or the errors
not fixed by the application yet). Even minor changes to the main error
message can break these filters. This makes it harder to improve
error messages. The key-values section is more structured and more
stable which makes it more suitable for filtering purposes.

The key-values section is not used to improve the main error message.
The latter should contain all the necessary information to understand
the error and present it in the most comprehensible way possible.