layers: Do not create wait timepoint after signal retired
This is follow-up to
8a41fe2 - Fix empty timeline assumption in Semaphore::EnqueueWait

Wait timepoint should not be registered when signal is already retired.
In this case wait should be resolved againt completed_ state.

There is assert in Semaphore::CanBinaryBeSignaled that checks that
binary timepoint has at least signal op. The new tests will trigger this
assert in case of a regression.