tests: Fix massive syncval tests failure on Windows Arm
The problem was that after DestroyedUnusedDescriptors all tests
started to fail. Investigation shown Adreno driver does not like
when uniform buffer is removed even it is not dynamically used.

The solution is simple. This test tests record time validation of
descriptors (ValidateDispatchDrawDescriptorSet). If we remove
QueueSubmit we still run the same validation but also make the test
stable on Adreno driver.