gpuav: Fix selection of inst desc set pipe layout
Added a test that showed issues with the selection of the
pipeline layout used to bind the instrumentation descriptor set.
Now, when trying to get pipeline layout from last bound pipeline,
if it has been destroyed it is temporarily re-created and used to bind
instrumentation descriptor set.
Trying to get a pipeline layout from another source (for the new test,
pipeline layout specified in last vkCmdPushConstants call) lead to
the dreaded vkCmdDraw-None-08600