state: Tear down state maps correctly in destructor
We had logic in the vkDestroyDevice() code path that cleans up the
state maps in a specific order to handle interdependencies between
state object types. This also needs to be called in the destructor,
in case an application exits without destroying vulkan objects. In
that case, everything gets destroyed via an atexit() callback and
it still needs to use the correct ordering.