tests: Reworked a bit driver info printing logic
A funny issue showed up while investigating issue #9109
When ran on his own,
NegativeDeviceFeatureProperty.PhysicalDeviceFeatures2
would fail (expected error not emitted), but would succeed when running
the whole NegativeDeviceFeatureProperty test suite.
I traced back the issue to the driver info printing logic: for this
info to be printed, the VK_KHR_get_physical_device_properties2 ext is
needed. It was automatically added when trying to print the driver
info, but only if said info had not already been printed. This explains
the issue I was seeing.
I decided to just stop adding this extension under the hood. Caveat is
if running a 1.0 test that does not add this extension, driver info
will now not be printed. I personnally don't like when extensions or
features are activated under the hood, to avoid the type of problems
described above. But being to print driver info no matter what is
convenient, so I will abide by team ruling on that matter.

Also, the "print driver info logic" just works out of luck. Not much
value in spending the time to do it properly, so I won't.
I described the problem in a new comment, inlined here:

The logic behind "printing the driver info only once"
works out of luck. The test suite implicitly relies on the fact that
VVL shared library loads and unloads (between tests) happen
in the same memory region, preserving statically initialised values.
At least simple values like "driver_printed",
that do not get destroyed or memset to 0.